UFS/FFS: Migrate to modern uintXX_t from u_intXX_t.
As per https://lists.freebsd.org/archives/freebsd-scsi/2023-July/000257.html
move to the modern uintXX_t. While here also migrate u_char to uint8_t.
Where other kernel interfaces allow, migrate u_long to uint64_t.

No functional changes intended.

Sponsored-by: The FreeBSD Foundation
(cherry picked from commit 831b1ff7913fb0b317a258a43dce5abefee03a31)

+++ sys/ufs/ufs/ufs_vfsops.c
@@ -115,7 +115,7 @@ ufs_quotactl(struct mount *mp, int cmds, uid_t id, void *arg)
 			return (EINVAL);
 		}
 	}
-	if ((u_int)type >= MAXQUOTAS) {
+	if ((uint64_t)type >= MAXQUOTAS) {
 		if (cmd == Q_QUOTAON || cmd == Q_QUOTAOFF)
 			vfs_unbusy(mp);
 		return (EINVAL);
